requirements.txt
Flask==3.0.3
Authlib==1.3.0
regex==2024.5.15
python-dotenv==1.0.1
pytest==8.2.1
requests==2.32.2